+++ NEWS.md
@@ -1,7 +1,9 @@
 # `CAST` 0.3.1
 
-* bugfix: CreateSpaceTimeFolds accepts tibbles
-* bugfix: CreateSpaceTimeFolds automatically reduces k if necessary
+* bugfix: 
+  * CreateSpaceTimeFolds accepts tibbles
+  * CreateSpaceTimeFolds automatically reduces k if necessary
+  * ffs accepts further arguments taken by caret::train
 * new feature: plot_ffs has option to plot selected variables only
 
 # `CAST` 0.3.0