+++ NEWS.md
@@ -1,3 +1,6 @@
+# `CAST` 0.3.2
+* new feature: "minVar": Instead of always starting with 2-pair combinations, ffs can now also be started with combinations of more variables (e.g starting with all combinations of 3)
+
 # `CAST` 0.3.1
 
 * bugfix: